Fix seccomp TestLogAct.
This change fixes TestLogAct test. This test was written in such as
way that once the filter was loaded, it blocked almost all system
calls, thereby making disabling the filter impossible and sometimes
causing the Go runtime to fail to allocate memory.

This fix simplifies the test and fixes these isuses.